requirements.txt
streamlit==1.32.0
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
accelerate==0.20.3
numpy==1.24.3
pandas==2.0.2
openpyxl==3.1.2
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
Pillow==9.5.0
opencv-python-headless==4.7.0.72
tqdm==4.65.0